requirements.txt
asgiref==3.8.1
Django==5.1
django-cors-headers==4.4.0
djangorestframework==3.15.2
filelock==3.15.4
fsspec==2024.6.1
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mpmath==1.3.0
networkx==3.3
numpy==1.26.4
pillow==10.4.0
setuptools==72.1.0
sqlparse==0.5.1
sympy==1.13.2
tbb==2021.13.1
torch==2.3.1
torchaudio==2.3.1
torchinfo==1.8.0
torchvision==0.18.1
typing_extensions==4.12.2
tzdata==2024.1
